core: move internet speed test to pardal
