+++ core/frontend/src/store/helper.ts
@@ -7,7 +7,7 @@ import Notifier from '@/libs/notifier'
 import { OneMoreTime } from '@/one-more-time'
 import store from '@/store'
 import { helper_service } from '@/types/frontend_services'
-import { InternetConnectionState, Service, SpeedTestResult } from '@/types/helper'
+import { InternetConnectionState, Service } from '@/types/helper'
 import back_axios, { isBackendOffline } from '@/utils/api'
 
 const notifier = new Notifier(helper_service)
@@ -96,58 +96,6 @@ class PingStore extends VuexModule {
         this.setHasInternet(InternetConnectionState.UNKNOWN)
         this.setReachableHosts([])
         notifier.pushBackError('INTERNET_CHECK_FAIL', error)
-      })
-  }
-
-  @Action
-  async checkInternetBestServer(): Promise<SpeedTestResult | void> {
-    return back_axios({
-      method: 'get',
-      url: `${this.API_URL}/internet_best_server`,
-      timeout: 15000,
-    })
-      .then((response) => response.data as SpeedTestResult)
-      .catch((error) => {
-        notifier.pushBackError('INTERNET_BEST_SERVER_FAIL', error)
-      })
-  }
-
-  @Action
-  async checkInternetDownloadSpeed(): Promise<SpeedTestResult | void> {
-    return back_axios({
-      method: 'get',
-      url: `${this.API_URL}/internet_download_speed`,
-      timeout: 15000,
-    })
-      .then((response) => response.data as SpeedTestResult)
-      .catch((error) => {
-        notifier.pushBackError('INTERNET_DOWNLOAD_SPEED_FAIL', error)
-      })
-  }
-
-  @Action
-  async checkInternetUploadSpeed(): Promise<SpeedTestResult | void> {
-    return back_axios({
-      method: 'get',
-      url: `${this.API_URL}/internet_upload_speed`,
-      timeout: 15000,
-    })
-      .then((response) => response.data as SpeedTestResult)
-      .catch((error) => {
-        notifier.pushBackError('INTERNET_UPLOAD_SPEED_FAIL', error)
-      })
-  }
-
-  @Action
-  async checkPreviousInternetTestResult(): Promise<SpeedTestResult | void> {
-    return back_axios({
-      method: 'get',
-      url: `${this.API_URL}/internet_test_previous_result`,
-      timeout: 15000,
-    })
-      .then((response) => response.data as SpeedTestResult)
-      .catch((error) => {
-        notifier.pushBackError('INTERNET_RESULT_SPEED_FAIL', error)
       })
   }
