+++ core/services/helper/main.py
@@ -9,7 +9,6 @@ import re
 import socket
 import subprocess
 from concurrent import futures
-from datetime import datetime
 from enum import Enum
 from functools import cache
 from io import BytesIO
@@ -36,13 +35,11 @@ from fastapi.responses import HTMLResponse
 from fastapi_versioning import VersionedFastAPI, version
 from loguru import logger
 from pydantic import BaseModel
-from speedtest import Speedtest
 from uvicorn import Config, Server
 
 from nginx_parser import parse_nginx_file
 
 SERVICE_NAME = "helper"
-SPEED_TEST: Optional[Speedtest] = None
 
 logging.basicConfig(handlers=[InterceptHandler()], level=logging.DEBUG)
 try:
@@ -51,12 +48,6 @@ except Exception as logger_e:
     print(f"Error: unable to set logger path: {logger_e}")
 
 logger.info("Starting Helper")
-
-try:
-    SPEED_TEST = Speedtest(secure=True)
-except Exception:
-    # When starting, the system may not be connected to the internet
-    pass
 
 
 class Website(Enum):
@@ -134,45 +125,6 @@ class ServiceInfo(BaseModel):
         if self.port not in Helper.attempts_left:
             return False
         return bool(Helper.attempts_left[self.port] <= 0)
-
-
-class SpeedtestServer(BaseModel):
-    url: str
-    lat: str
-    lon: str
-    name: str
-    country: str
-    cc: str
-    sponsor: str
-    id: str
-    host: str
-    d: float
-    latency: float
-
-
-class SpeedtestClient(BaseModel):
-    ip: str
-    lat: str
-    lon: str
-    isp: str
-    isprating: str
-    rating: str
-    ispdlavg: str
-    ispulavg: str
-    loggedin: str
-    country: str
-
-
-class SpeedTestResult(BaseModel):
-    download: float
-    upload: float
-    ping: float
-    server: SpeedtestServer
-    timestamp: datetime
-    bytes_sent: int
-    bytes_received: int
-    share: Optional[str] = None
-    client: SpeedtestClient
 
 
 class SimpleHttpResponse(BaseModel):
@@ -598,59 +550,6 @@ def software_id() -> Any:
         return str(uuid_obj)
     except Exception as exception:
         raise HTTPException(status_code=400, detail="Error: {exception}") from exception
-
-
-@fast_api_app.get(
-    "/internet_best_server",
-    response_model=SpeedTestResult,
-    summary="Check internet best server for test from BlueOS.",
-)
-@version(1, 0)
-async def internet_best_server(interface_addr: Optional[str] = None) -> Any:
-    # Since we are finding a new server, clear previous results
-    # pylint: disable=global-statement
-    global SPEED_TEST
-    SPEED_TEST = Speedtest(secure=True, source_address=interface_addr)
-    SPEED_TEST.get_best_server()
-    return SPEED_TEST.results.dict()
-
-
-@fast_api_app.get(
-    "/internet_download_speed",
-    response_model=SpeedTestResult,
-    summary="Check internet download speed test from BlueOS.",
-)
-@version(1, 0)
-async def internet_download_speed() -> Any:
-    if not SPEED_TEST:
-        raise RuntimeError("SPEED_TEST not initialized, initialize server search.")
-    SPEED_TEST.download()
-    return SPEED_TEST.results.dict()
-
-
-@fast_api_app.get(
-    "/internet_upload_speed",
-    response_model=SpeedTestResult,
-    summary="Check internet upload speed test from BlueOS.",
-)
-@version(1, 0)
-async def internet_upload_speed() -> Any:
-    if not SPEED_TEST:
-        raise RuntimeError("SPEED_TEST not initialized, initialize server search.")
-    SPEED_TEST.upload(pre_allocate=False)
-    return SPEED_TEST.results.dict()
-
-
-@fast_api_app.get(
-    "/internet_test_previous_result",
-    response_model=SpeedTestResult,
-    summary="Return previous result of internet speed test.",
-)
-@version(1, 0)
-async def internet_test_previous_result() -> Any:
-    if not SPEED_TEST:
-        raise RuntimeError("SPEED_TEST not initialized, initialize server search.")
-    return SPEED_TEST.results.dict()
 
 
 @fast_api_app.get(
